chore(main): rebrand module to inventory from stock
affects: @commercetools/inventories-exporter

the ensures consistency with the API
