+++ docs/SUMMARY.md
@@ -26,3 +26,4 @@
   * [CSV Order Parser](/cli/csv-parser-orders.md)
   * [Discount Code Generator](/cli/discount-code-generator.md)
   * [Discount Code Importer](/cli/discount-code-importer.md)
+  * [Inventories Exporter](/cli/inventories-exporter.md)
