+++ packages/inventories-exporter/bin/inventories-exporter.js
