+++ packages/inventories-exporter/package.json
@@ -4,7 +4,8 @@
   "description": "Export inventories from the commercetools platform.",
   "keywords": [
     "commercetools",
-    "discount",
+    "inventories",
+    "stock",
     "code",
     "export",
     "exporter"
@@ -21,9 +22,30 @@
     "email": "abimbola.idowu@commercetools.de",
     "url": "hisabimbola.com"
   },
+  "bin": {
+    "inventoriesexporter": "bin/inventories-exporter.js"
+  },
   "main": "lib/index.js",
   "files": [
     "lib",
     "bin"
-  ]
+  ],
+  "dependencies": {
+    "@commercetools/api-request-builder": "^3.1.1",
+    "@commercetools/get-credentials": "^2.1.0",
+    "@commercetools/sdk-client": "^1.4.1",
+    "@commercetools/sdk-middleware-auth": "^3.1.0",
+    "@commercetools/sdk-middleware-http": "^2.1.1",
+    "@commercetools/sdk-middleware-user-agent": "^1.2.0",
+    "JSONStream": "^1.2.1",
+    "fast-csv": "^2.4.0",
+    "lodash.memoize": "^4.1.2",
+    "npmlog": "^4.1.2",
+    "pretty-error": "^2.1.1",
+    "yargs": "^8.0.2"
+  },
+  "devDependencies": {
+    "common-tags": "^1.4.0",
+    "streamtest": "^1.2.2"
+  }
 }
