--- packages/stock-exporter/package.json
@@ -1,50 +1,0 @@
-{
-  "name": "@commercetools/stock-exporter",
-  "version": "1.0.0-development",
-  "description": "Export discount codes from the commercetools platform.",
-  "keywords": [
-    "commercetools",
-    "discount",
-    "code",
-    "export",
-    "exporter"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/commercetools/nodejs/tree/master/packages/stock-exporter"
-  },
-  "homepage": "https://commercetools.github.io/nodejs/cli/stockExporter.html",
-  "bugs": "https://github.com/commercetools/nodejs/issues",
-  "license": "MIT",
-  "author": {
-    "name": "Abimbola Idowu",
-    "email": "abimbola.idowu@commercetools.de",
-    "url": "hisabimbola.com"
-  },
-  "bin": {
-    "stockexporter": "bin/stock-exporter.js"
-  },
-  "main": "lib/index.js",
-  "files": [
-    "lib",
-    "bin"
-  ],
-  "dependencies": {
-    "@commercetools/api-request-builder": "^3.1.1",
-    "@commercetools/get-credentials": "^2.1.0",
-    "@commercetools/sdk-client": "^1.4.1",
-    "@commercetools/sdk-middleware-auth": "^3.1.0",
-    "@commercetools/sdk-middleware-http": "^2.1.1",
-    "@commercetools/sdk-middleware-user-agent": "^1.2.0",
-    "JSONStream": "^1.2.1",
-    "fast-csv": "^2.4.0",
-    "lodash.memoize": "^4.1.2",
-    "npmlog": "^4.1.2",
-    "pretty-error": "^2.1.1",
-    "yargs": "^8.0.2"
-  },
-  "devDependencies": {
-    "common-tags": "^1.4.0",
-    "streamtest": "^1.2.2"
-  }
-}
