--- packages/stock-exporter/readme.md
@@ -1,11 +1,0 @@
-# stock-exporter
-
-Export inventories(stock) from the commercetools platform.
-
-more info here: https://commercetools.github.io/nodejs/api/stock-exporter/
-
-## Install
-
-```bash
-npm install --save @commercetools/stock-exporter
-```
